Working switch2led demonstration on tic55x
This is a first release that is starting to work on real
phone hardware: Grandstream BT200.  The part that works
is the backlight to the LCD and the hook contact, but
that at least shows control.  Work has been done on the
timer actions, but that has not been made to work yet,
at least not based on interrupts.

+++ include/0cpm/netcmd.h
@@ -56,7 +56,7 @@
  * }
  */
 
-#define BPF_BEGIN(fname) uint32_t fname (uint8_t *P, uint32_t L, uint32_t *M) { register uint32_t A, X=htonl ((uint32_t) P), S; register int k;
+#define BPF_BEGIN(fname) intptr_t fname (uint8_t *P, uint32_t L, uint32_t *M) { register uint32_t A, X=htonl ((uint32_t) P), S; register int k;
 #define BPF_LABEL(l) l:
 #define BPF_END() }
 #define BPF_TAIL(fname) return (fname) (P, L, M);
